#66 Moved Core entities into Core namespace
