+++ src/AdminAssistant.DAL/EntityFramework/Model/Budget/BudgetEntity.cs
@@ -7,6 +7,6 @@ namespace AdminAssistant.DAL.EntityFramework.Model.Budget
         public int OwnerID { get; internal set; }
         public string BudgetName { get; internal set; } = string.Empty;
 
-        public AuditEntity Audit { get; internal set; } = null!;
+        public Core.AuditEntity Audit { get; internal set; } = null!;
     }
 }
