+++ src/AdminAssistant.DAL/EntityFramework/Model/Contacts/AddressEntity.cs
@@ -5,6 +5,6 @@ namespace AdminAssistant.DAL.EntityFramework.Model.Contacts
         public int AddressID { get; set; }
         public int AuditID { get; internal set; }
 
-        public AuditEntity Audit { get; internal set; } = null!;
+        public Core.AuditEntity Audit { get; internal set; } = null!;
     }
 }
