+++ src/AdminAssistant.DAL/EntityFramework/Model/Contacts/ContactEntity.cs
@@ -10,6 +10,6 @@ namespace AdminAssistant.DAL.EntityFramework.Model.Contacts
         public DateTime DateOfBirth { get; set; }
         public int AuditID { get; internal set; }
 
-        public AuditEntity Audit { get; internal set; } = null!;
+        public Core.AuditEntity Audit { get; internal set; } = null!;
     }
 }
