+++ src/AdminAssistant.DAL/EntityFramework/Model/Core/AuditEntity.cs
@@ -3,7 +3,7 @@ using AdminAssistant.DAL.EntityFramework.Model.Accounts;
 using AdminAssistant.DAL.EntityFramework.Model.AssetRegister;
 using AdminAssistant.DAL.EntityFramework.Model.Contacts;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Core
 {
     public class AuditEntity
     {
