+++ src/AdminAssistant.DAL/EntityFramework/Model/Core/OwnerEntity.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using AdminAssistant.DAL.EntityFramework.Model.Accounts;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Core
 {
     public class OwnerEntity
     {
