+++ src/AdminAssistant.DAL/EntityFramework/Model/Core/PermissionEntity.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Core
 {
     public class PermissionEntity
     {
