+++ src/AdminAssistant.DAL/EntityFramework/Model/Core/SettingEntity.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace AdminAssistant.DAL.EntityFramework.Model
+namespace AdminAssistant.DAL.EntityFramework.Model.Core
 {
     public class SettingEntity
     {
