+++ src/AdminAssistant.DAL/EntityFramework/Model/Documents/DocumentEntity.cs
@@ -6,6 +6,6 @@ namespace AdminAssistant.DAL.EntityFramework.Model.Documents
         public int AuditID { get; internal set; }
         public int OwnerID { get; internal set; }
 
-        public AuditEntity Audit { get; internal set; } = null!;
+        public Core.AuditEntity Audit { get; internal set; } = null!;
     }
 }
